+++ saessak-front/src/components/admin/AdminImage.jsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+
+// 한페이지에 60개
+// 어드민 체크한 이미지는 보이기/제외 버튼 넣기
+// 최신이미지 기준으로 로딩시작 / 
+// 무한스크롤로 페이징처리하기
+// 저장버튼은 fixed 로 따라다니기
+
+
+const AdminImage = ({setModalData, page}) => {
+  return (
+    <div>
+    </div>
+  )
+}
+
+export default AdminImage
